Move files to new locations and fixup __init__.py files
