+++ scanpointgenerator/__init__.py
@@ -1,12 +1,4 @@
-from scanpointgenerator.point import Point  # noqa
-from scanpointgenerator.generator import Generator  # noqa
-from scanpointgenerator.arraygenerator import ArrayGenerator  # noqa
-from scanpointgenerator.compoundgenerator import CompoundGenerator  # noqa
-from scanpointgenerator.linegenerator import LineGenerator  # noqa
-from scanpointgenerator.lissajousgenerator import LissajousGenerator  # noqa
-from scanpointgenerator.mutator import Mutator  # noqa
-from scanpointgenerator.randomoffsetmutator import RandomOffsetMutator  # noqa
-from scanpointgenerator.spiralgenerator import SpiralGenerator  # noqa
-from scanpointgenerator.excluder import Excluder  # noqa
-from scanpointgenerator.plotgenerator import plot_generator  # noqa
-from scanpointgenerator.compat import range_  # noqa
+from scanpointgenerator.core import *
+from scanpointgenerator.generators import *
+import scanpointgenerator.rois
+import scanpointgenerator.mutators
