+++ scanpointgenerator/generators/__init__.py
@@ -1,0 +1,4 @@
+from scanpointgenerator.generators.arraygenerator import ArrayGenerator
+from scanpointgenerator.generators.linegenerator import LineGenerator
+from scanpointgenerator.generators.lissajousgenerator import LissajousGenerator
+from scanpointgenerator.generators.spiralgenerator import SpiralGenerator
