+++ scanpointgenerator/mutators/__init__.py
@@ -1,0 +1,2 @@
+from scanpointgenerator.mutators.fixeddurationmutator import FixedDurationMutator
+from scanpointgenerator.mutators.randomoffsetmutator import RandomOffsetMutator
