+++ scanpointgenerator/rois/__init__.py
@@ -1,0 +1,7 @@
+from scanpointgenerator.rois.circular_roi import CircularROI
+from scanpointgenerator.rois.elliptical_roi import EllipticalROI
+from scanpointgenerator.rois.linear_roi import LinearROI
+from scanpointgenerator.rois.point_roi import PointROI
+from scanpointgenerator.rois.polygonal_roi import PolygonalROI
+from scanpointgenerator.rois.rectangular_roi import RectangularROI
+from scanpointgenerator.rois.sector_roi import SectorROI
