+++ scanpointgenerator/rois/rectangular_roi.py
@@ -1,6 +1,6 @@
 from math import cos, sin
 
-from scanpointgenerator.roi import ROI
+from scanpointgenerator.core import ROI
 
 
 @ROI.register_subclass("scanpointgenerator:roi/RectangularROI:1.0")
